+++ IntCodeCpu/Instructions/InstrMultiply.cs
@@ -14,9 +14,9 @@ namespace IntCodeCpu.Instructions
 
     public override eProcState Execute(Processor proc, Input input, Output output, InstructionParams instructParams)
     {
-      int value1 = GetParam(0, proc, instructParams);
-      int value2 = GetParam(1, proc, instructParams);
-      int res = value1 * value2;
+      long value1 = GetParam(0, proc, instructParams);
+      long value2 = GetParam(1, proc, instructParams);
+      long res = value1 * value2;
       SetParam(2, res, proc, instructParams);
       proc.InstructionPointer += 4;
       return eProcState.Continue;
