+++ IntCodeCpu/Instructions/InstructionParams.cs
@@ -8,9 +8,9 @@ namespace IntCodeCpu
 {
   public class InstructionParams
   {
-    public int OpCode { get; set; }
+    public long OpCode { get; set; }
     public List<int> ParamsMode { get; set; } = new List<int>();
-    public List<int> Values { get; set; } = new List<int>();
+    public List<long> Values { get; set; } = new List<long>();
     public InstructionParams()
     {
     }
